Tests/LibWeb/Text/expected/Worker/Worker-referrer.txt

Got referrer: http://127.0.0.1/worker-referrer/worker.js
